+++ hazelcast/src/main/java/com/hazelcast/map/client/MapTryRemoveRequest.java
@@ -38,13 +38,13 @@ public class MapTryRemoveRequest extends KeyBasedClientRequest implements Portab
 
     protected String name;
     protected Data key;
-    protected int threadId;
+    protected long threadId;
     protected long timeout;
 
     public MapTryRemoveRequest() {
     }
 
-    public MapTryRemoveRequest(String name, Data key, int threadId, long timeout) {
+    public MapTryRemoveRequest(String name, Data key, long threadId, long timeout) {
         this.name = name;
         this.key = key;
         this.threadId = threadId;
@@ -77,7 +77,7 @@ public class MapTryRemoveRequest extends KeyBasedClientRequest implements Portab
 
     public void write(PortableWriter writer) throws IOException {
         writer.writeUTF("n", name);
-        writer.writeInt("t", threadId);
+        writer.writeLong("t", threadId);
         writer.writeLong("timeout", timeout);
         final ObjectDataOutput out = writer.getRawDataOutput();
         key.writeData(out);
@@ -85,7 +85,7 @@ public class MapTryRemoveRequest extends KeyBasedClientRequest implements Portab
 
     public void read(PortableReader reader) throws IOException {
         name = reader.readUTF("n");
-        threadId = reader.readInt("t");
+        threadId = reader.readLong("t");
         timeout = reader.readLong("timeout");
         final ObjectDataInput in = reader.getRawDataInput();
         key = new Data();
