+++ hazelcast/src/main/java/com/hazelcast/util/ThreadUtil.java
@@ -18,20 +18,20 @@ package com.hazelcast.util;
 
 public final class ThreadUtil {
 
-    private static final ThreadLocal<Integer> threadLocal = new ThreadLocal<Integer>();
+    private static final ThreadLocal<Long> threadLocal = new ThreadLocal<Long>();
 
     private ThreadUtil(){}
 
-    public static int getThreadId() {
-        final Integer threadId = threadLocal.get();
+    public static long getThreadId() {
+        final Long threadId = threadLocal.get();
         if (threadId != null) {
             return threadId;
         }
-        return (int) Thread.currentThread().getId();  // TODO: @mm - thread-id is truncated from native thread id
+        return Thread.currentThread().getId();
     }
 
     public static void setThreadId(long threadId) {
-        threadLocal.set((int)threadId);
+        threadLocal.set(threadId);
     }
 
     public static void removeThreadId() {
